# cmd/kubectl-audit-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-cluster aggregation server for kubectl-audit: clusters push findings.json scans here, which are deduplicated and stored for cross-cluster triage, centralized knowledge base/exclusion rules, and automation rule evaluation.",
        "title": "kubectl-audit-server API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/automation-rules": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Lists every automation rule.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "automation-rules"
                ],
                "summary": "List automation rules",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/internal_server_http.automationRuleDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid admin token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Creates a rule that fires an action (e.g. notify/file-jira) when its trigger matches a finding during evaluation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "automation-rules"
                ],
                "summary": "Create an automation rule",
                "parameters": [
                    {
                        "description": "Automation rule to create",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server_http.automationRuleDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_server_http.automationRuleDTO"
                        }
                    },
                    "400": {
                        "description": "invalid request body / missing name or action.type",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid admin token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/automation-rules/evaluate": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Evaluates every enabled automation rule against current findings immediately, instead of waiting for the periodic ticker. Useful to demo automation without waiting for a timer.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "automation-rules"
                ],
                "summary": "Run one automation evaluation pass",
                "responses": {
                    "200": {
                        "description": "results",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/internal_server_http.automationEvaluationResultDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid admin token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "automation runner not configured / evaluation failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/automation-rules/{id}": {
            "delete": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Deletes one automation rule by ID.",
                "tags": [
                    "automation-rules"
                ],
                "summary": "Delete an automation rule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Automation rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "deleted"
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid admin token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "no such automation rule",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Replaces the name/enabled/trigger/action of an existing automation rule. ID and createdAt are preserved from the existing record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "automation-rules"
                ],
                "summary": "Update an automation rule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Automation rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New rule content",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server_http.automationRuleDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server_http.automationRuleDTO"
                        }
                    },
                    "400": {
                        "description": "invalid id / invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid admin token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "no such automation rule",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_server_http.automationActionDTO": {
            "type": "object",
            "properties": {
                "prompt": {
                    "type": "string"
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "internal_server_http.automationEvaluationResultDTO": {
            "type": "object",
            "properties": {
                "attempted": {
                    "type": "boolean"
                },
                "clusterId": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "fingerprint": {
                    "type": "string"
                },
                "ruleId": {
                    "type": "string"
                },
                "ruleName": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "internal_server_http.automationRuleDTO": {
            "type": "object",
            "properties": {
                "action": {
                    "$ref": "#/definitions/internal_server_http.automationActionDTO"
                },
                "createdAt": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "trigger": {
                    "$ref": "#/definitions/internal_server_http.automationTriggerDTO"
                }
            }
        },
        "internal_server_http.automationTriggerDTO": {
            "type": "object",
            "properties": {
                "minSeverity": {
                    "type": "string"
                },
                "noJiraLinkForHours": {
                    "type": "integer"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin bearer token (env ADMIN_TOKEN). Send as \"Bearer \u003ctoken\u003e\". Required for cluster registration and every organization-level config endpoint (knowledge base, exclusion rules, automation rules, triage).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ClusterAuth": {
            "description": "Per-cluster bearer token minted by POST /clusters. Send as \"Bearer \u003ctoken\u003e\". Required for ingest endpoints only — scoped to that cluster's own data, and does not grant triage read/write (see AdminAuth).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}